requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.1
dnspython==2.4.2
docx2txt==0.8
docxpy==0.8.5
email-validator==2.0.0.post2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
fsspec==2023.9.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.306
langcodes==3.3.0
langsmith==0.0.41
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
orjson==3.9.7
packaging==23.2
pandas==2.1.1
pathy==0.10.2
pgvector==0.2.3
preshed==3.0.9
psycopg2-binary==2.9.8
pydantic==1.10.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
starlette==0.27.0
tenacity==8.2.3
thinc==8.2.1
tiktoken==0.5.1
tqdm==4.66.1
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.17.0
wasabi==1.1.2
watchfiles==0.20.0
weasel==0.3.1
websockets==11.0.3
wget==3.2
yarl==1.9.2